I split more concerns, added types/tsconfig, and updated docs while keeping behavior unchanged.
What I changed
- Utilities: Extracted DOM helpers.
  - src/utils/legacy.ts: `hideLegacyControls(ids)` for hiding old form elements.
  - src/utils/mount.ts: `mountIfPresent(id, node)` for safe mounting.
- Types: Centralized user shape.
  - src/types.ts: `User` type.
- Components: Kept islands modular.
  - src/components/SetupForm.tsx
  - src/components/ManagementSettings.tsx
  - src/components/DndManager.tsx
- Entry Module: Slimmer entry with clear responsibilities.
  - src/atlaskit-forms.tsx now imports components/utils, calls `hideLegacyControls`, mounts islands, exposes only `init` on `window.FormsMount` (same runtime behavior).
- Tooling/Docs: Added type config and clarified structure.
  - tsconfig.json with strict, noEmit, JSX react-jsx.
  - README.md and docs/architecture-ja.md now reflect the new component/util layout.

Why this helps
- Clear separation of concerns: Setup, Settings, and DnD logic live in focused files.
- Reusable utilities: Common DOM behaviors (mount/hide legacy) are shared.
- Safer globals: Single, explicit export of `window.FormsMount = { init }` avoids accidental exposure.
- Type safety: A central `User` type and strict TS config improve maintainability.

Behavior preserved
- Build entry remains `src/atlaskit-forms.tsx` producing the same IIFE file.
- Only `init` is exposed on `window.FormsMount`, matching current behavior.
- All DOM interop and window hooks are unchanged.

Verification notes
- Deps are not installed in this sandbox; building will error until `pnpm install` runs.
- Local steps:
  - pnpm install
  - export SUPABASE_URL=... && export SUPABASE_ANON_KEY=...
  - pnpm run build
  - Open `public/index.html`, confirm:
    - Setup form mounts.
    - Legacy inputs are hidden, but values stay in sync.
    - Legacy DnD still active (Atlaskit DnD not globally exposed by design).

Next refinements (optional)
- Expose optional islands: If you want to switch the UI from legacy Sortable to Atlaskit, I can expose `initDnd` and `mountManagement` on `window.FormsMount` to let index.html mount them. This changes behavior, so I held off.
- Add linting: ESLint + Prettier config and a `pnpm run lint` script.
- Add `typecheck` script: `tsc --noEmit` to CI.
- Unit tests: Smoke tests for interop and DnD constraints.

Would you like me to expose `initDnd` and `mountManagement` globally now, or proceed with linting setup first?
